refactor: 가짜 객체 제거 및 SpringBootTest로 변경
