+++ src/test/java/wooteco/subway/dao/JdbcLineDaoTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao.line;
+package wooteco.subway.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
+import wooteco.subway.dao.line.JdbcLineDao;
+import wooteco.subway.dao.line.LineDao;
 import wooteco.subway.domain.Line;
 
 @JdbcTest
