+++ src/test/resources/init.sql
@@ -1,0 +1,7 @@
+ALTER TABLE STATION
+    ALTER COLUMN id RESTART WITH 1;
+ALTER TABLE LINE
+    ALTER COLUMN id RESTART WITH 1;
+ALTER TABLE SECTION
+    ALTER COLUMN id RESTART WITH 1;
+
